execution: relocate the app-debug simulator from cli
The mini-app simulator moves from `packages/cli/src/app-debug/` to
`packages/execution/src/app-debug/`, exported as
`@nodetool-ai/execution/app-debug` the way `execution/debug` already is, so
the agent build loop (M2) and the server can simulate an app without
depending on the CLI.

`simulateApp` takes a `ResolvedAppTarget` and an injected workflow runner and
returns the `AppDebugReport`. It reads no database and writes no files: the
CLI keeps target resolution (`app-target.ts`), the default output directory,
and bundle writing, and passes the runner plus a sink for each run's raw
message stream. `runAppDebug` keeps its signature, still loads the kernel
server runner lazily on the first run, and `app-debug/index.ts` re-exports the
moved surface.

The target and server-run vocabulary the simulator shares with the workflow
harness (`DebugGraph`, `DebugTargetInfo`, `TraceSummary`, `ServerRunReport`,
`DebugVerdict`) moves to `execution/debug/types.ts`; the CLI's `debug/types.ts`
re-exports it.

Behavior-preserving: `scripts/app-debug-parity.mjs` captures
`app debug --json` over all 11 shipped example bundles, static and running,
before and after. The reports are byte-identical except research-desk, whose
workflow fetches a live RSS feed; with the feed payloads projected out that
one matches too.
