+++ packages/cli/src/app-debug/index.ts
@@ -1,32 +1,29 @@
+/**
+ * The CLI's app-debug surface: target resolution plus the bundle-writing
+ * harness, over the simulator in `@nodetool-ai/execution/app-debug`. The
+ * simulator's exports are re-exported here so importers keep one entry point.
+ */
 export { runAppDebug, defaultInteractions } from "./harness.js";
 export type { AppDebugDeps } from "./harness.js";
-export {
-  parseAppSpec,
-  parseAppDocument,
-  documentOperations,
-  operationSpec,
-  extractAppIO,
-  validateApp,
-  bindingScopeFor
-} from "./app-spec.js";
-export type { AppContext } from "./app-spec.js";
 export { resolveAppTarget } from "./app-target.js";
 export type {
   AppApplicationRecord,
   AppTargetDeps,
-  AppWorkflowRecord,
-  ResolvedAppTarget
+  AppWorkflowRecord
 } from "./app-target.js";
 export {
+  bindingScopeFor,
+  documentOperations,
+  effectiveTimeoutMs,
+  extractAppIO,
   HeadlessAppRuntime,
   InMemoryResourceProvider,
-  effectiveTimeoutMs,
-  seedResourceHint
-} from "./runtime.js";
-export type {
-  HeadlessResourceCommand,
-  HeadlessResourceProvider,
-  ResourceItem
-} from "./runtime.js";
-export { renderAppReportMarkdown } from "./markdown.js";
-export type * from "./types.js";
+  operationSpec,
+  parseAppDocument,
+  parseAppSpec,
+  renderAppReportMarkdown,
+  seedResourceHint,
+  simulateApp,
+  validateApp
+} from "@nodetool-ai/execution/app-debug";
+export type * from "@nodetool-ai/execution/app-debug";
