+++ packages/execution/src/app-debug/index.ts
@@ -1,0 +1,40 @@
+/**
+ * App-debug surface: the headless mini-app simulator.
+ *
+ * One simulation of an app document — widget spec, static wiring check,
+ * interaction script, run folding, verdict — for every host that reports on an
+ * app: the `nodetool app debug` harness, the agent-facing build loop, and the
+ * server. Target resolution and bundle writing stay with the host; nothing here
+ * reads a database or a file.
+ */
+export { simulateApp, defaultInteractions } from "./simulate.js";
+export type {
+  AppServerRunInput,
+  AppServerRunOutcome,
+  AppSimulationDeps
+} from "./simulate.js";
+export {
+  parseAppSpec,
+  parseAppDocument,
+  documentOperations,
+  operationSpec,
+  extractAppIO,
+  validateApp,
+  bindingScopeFor
+} from "./app-spec.js";
+export type { AppContext } from "./app-spec.js";
+export {
+  HeadlessAppRuntime,
+  InMemoryResourceProvider,
+  effectiveTimeoutMs,
+  seedResourceHint
+} from "./runtime.js";
+export type {
+  HeadlessOperationInit,
+  HeadlessResourceCommand,
+  HeadlessResourceProvider,
+  HeadlessRunResult,
+  ResourceItem
+} from "./runtime.js";
+export { renderAppReportMarkdown } from "./markdown.js";
+export type * from "./types.js";
