+++ packages/execution/src/app-debug/simulate.ts
@@ -1,0 +1,859 @@
+/**
+ * The app simulator: runs a resolved mini app headlessly.
+ *
+ * Parses the app document into a widget spec and a set of operations,
+ * statically validates the wiring, then simulates the app the way the web and
+ * mobile runtimes do: seed input and variable defaults, apply params, attach
+ * the resource collections the script seeds, execute the interaction script
+ * (each `run` action is a full workflow run on the injected server runner,
+ * subject to the operation's policy and timeout), fold the message stream into
+ * the app's reactive values, and re-evaluate every widget's conditions — a step
+ * no user could perform, because the widget is hidden or disabled, fails.
+ *
+ * Nothing here touches a database or a filesystem. The host resolves the target
+ * (`ResolvedAppTarget`), supplies the workflow runner, and writes whatever
+ * bundle it wants; the returned `AppDebugReport` is the whole result.
+ */
+import type { ProcessingMessage } from "@nodetool-ai/protocol";
+import type {
+  DebugGraph,
+  DebugVerdict,
+  ServerRunReport
+} from "../debug/types.js";
+import { previewValue } from "../debug/collector.js";
+import {
+  DEFAULT_OPERATION_ID,
+  eventToAction,
+  parseBinding,
+  resolveBinding,
+  stateKey,
+  type OperationBinding
+} from "@nodetool-ai/app-runtime";
+import {
+  bindingScopeFor,
+  documentOperations,
+  extractAppIO,
+  operationSpec,
+  parseAppSpec,
+  validateApp,
+  type AppContext
+} from "./app-spec.js";
+import {
+  effectiveTimeoutMs,
+  HeadlessAppRuntime,
+  type HeadlessOperationInit
+} from "./runtime.js";
+import type {
+  AppDebugOptions,
+  AppDebugReport,
+  AppIO,
+  AppSpec,
+  AppWidgetSpec,
+  InteractionRecord,
+  InteractionStep,
+  ResolvedAppTarget,
+  SeedResourceItem
+} from "./types.js";
+
+/** What the simulator hands a host's workflow runner. */
+export interface AppServerRunInput {
+  graph: DebugGraph;
+  workflowId: string | null;
+  params: Record<string, unknown>;
+  timeoutMs?: number;
+}
+
+/** What it needs back: the run's report and the raw stream to fold. */
+export interface AppServerRunOutcome {
+  report: ServerRunReport;
+  rawMessages: ProcessingMessage[];
+}
+
+export interface AppSimulationDeps {
+  /** Load a workflow by DB id — an operation may name one the target lacks. */
+  loadFromDb: (
+    id: string
+  ) => Promise<{ graph: DebugGraph; app_doc?: unknown } | null>;
+  /** Execute one operation's workflow. */
+  runOnServer: (input: AppServerRunInput) => Promise<AppServerRunOutcome>;
+  /** Progress/log sink. */
+  onLog?: (line: string) => void;
+  /**
+   * Persist one run's raw message stream, returning the path to record on the
+   * run report. Hosts that write no bundle omit it.
+   */
+  onRunMessages?: (
+    runIndex: number,
+    messages: ProcessingMessage[]
+  ) => Promise<string | null>;
+}
+
+/** Resolve a step's widget reference: exact id, unique type, or unique label. */
+function findWidget(spec: AppSpec, ref: string): AppWidgetSpec | string {
+  const byId = spec.widgets.find((w) => w.id === ref);
+  if (byId) return byId;
+  const byType = spec.widgets.filter((w) => w.type === ref);
+  if (byType.length === 1) return byType[0];
+  if (byType.length > 1) {
+    return `widget reference "${ref}" is ambiguous — ${byType.length} widgets of that type; use the component id.`;
+  }
+  const byLabel = spec.widgets.filter((w) => w.label === ref);
+  if (byLabel.length === 1) return byLabel[0];
+  if (byLabel.length > 1) {
+    return `widget reference "${ref}" is ambiguous — ${byLabel.length} widgets share that label; use the component id.`;
+  }
+  return `no widget matches "${ref}" (tried id, type, and label).`;
+}
+
+/**
+ * The app's natural run trigger, mirroring what a user would do first: click
+ * the first `click`→`run` widget (a Run button), else touch the first
+ * `change`→`run` write widget.
+ */
+export function defaultInteractions(spec: AppSpec): InteractionStep[] {
+  const clicker = spec.widgets.find((w) =>
+    w.events.some((e) => e.trigger === "click" && e.kind === "run")
+  );
+  if (clicker) return [{ click: clicker.id }];
+  const changer = spec.widgets.find(
+    (w) =>
+      w.bindingMode === "write" &&
+      w.events.some((e) => e.trigger === "change" && e.kind === "run")
+  );
+  if (changer) return [{ change: changer.id, value: undefined }];
+  return [];
+}
+
+function describeStep(step: InteractionStep): string {
+  if ("set" in step) return `set ${step.set.key}`;
+  if ("seedResource" in step) return `seedResource ${step.seedResource.id}`;
+  if ("click" in step) return `click ${step.click}`;
+  if ("run" in step) return `run ${step.run}`;
+  if ("cancel" in step) return `cancel ${step.cancel}`;
+  return `change ${step.change}`;
+}
+
+/**
+ * Node types that emit on one of several output handles per run. A widget fed
+ * from one of their branches is *expected* to stay empty whenever the other
+ * branch is taken, so a single run cannot tell "the untaken branch" from "a
+ * branch that can never be taken".
+ */
+const BRANCHING_NODE_TYPES = new Set(["nodetool.control.If"]);
+
+/**
+ * Node ids reachable only through a branching node. Walks forward from every
+ * branch point over the graph's edges.
+ */
+function conditionalNodeIds(graph: DebugGraph): Set<string> {
+  const downstream = new Map<string, string[]>();
+  for (const edge of graph.edges) {
+    const source = typeof edge.source === "string" ? edge.source : null;
+    const target = typeof edge.target === "string" ? edge.target : null;
+    if (!source || !target) continue;
+    const list = downstream.get(source);
+    if (list) list.push(target);
+    else downstream.set(source, [target]);
+  }
+
+  const conditional = new Set<string>();
+  const queue: string[] = [];
+  for (const node of graph.nodes) {
+    const id = typeof node.id === "string" ? node.id : null;
+    const type = typeof node.type === "string" ? node.type : null;
+    if (id && type && BRANCHING_NODE_TYPES.has(type)) queue.push(id);
+  }
+  while (queue.length > 0) {
+    const current = queue.pop() as string;
+    for (const next of downstream.get(current) ?? []) {
+      if (conditional.has(next)) continue;
+      conditional.add(next);
+      queue.push(next);
+    }
+  }
+  return conditional;
+}
+
+/** Node ids keyed by the `name` their data carries, for name-form bindings. */
+function nodeIdsByName(graph: DebugGraph): Map<string, string> {
+  const byName = new Map<string, string>();
+  for (const node of graph.nodes) {
+    const id = typeof node.id === "string" ? node.id : null;
+    // Runner shape carries node props under `properties`; editor JSON uses
+    // `data`, and this runs against both.
+    const props = (node.properties ?? node.data) as
+      | Record<string, unknown>
+      | undefined;
+    const name = typeof props?.name === "string" ? props.name : null;
+    if (id && name && !byName.has(name)) byName.set(name, id);
+  }
+  return byName;
+}
+
+/**
+ * What a headless run leaves to the browser. Conditions and `format` are
+ * simulated, so only what needs a DOM (or a provider the simulator has not
+ * grown yet) belongs here.
+ */
+const NOT_SIMULATED: ReadonlyArray<string> = [
+  "Layout, styling, focus, and scroll — nothing here renders a DOM.",
+  "Stored resource collections — a run reads the seeded in-memory provider, never the database, and `openResource` has no editor to open.",
+  "Reactive subgraph runs — the browser reruns one input's downstream subgraph, the harness runs the whole workflow."
+];
+
+function buildAppVerdict(
+  report: Pick<
+    AppDebugReport,
+    | "validation"
+    | "interactions"
+    | "runs"
+    | "widgets"
+    | "spec"
+    | "invocations"
+    | "resources"
+  >,
+  ranWorkflow: boolean,
+  graph: DebugGraph,
+  conditionIssues: ReadonlyArray<string>
+): DebugVerdict {
+  const issues: string[] = [...report.validation.errors, ...conditionIssues];
+  const warnings: string[] = [];
+
+  for (const interaction of report.interactions) {
+    if (interaction.error) {
+      issues.push(`Interaction "${interaction.step}": ${interaction.error}`);
+    }
+  }
+  for (const run of report.runs) {
+    if (!run.ok) {
+      issues.push(`Run ended ${run.status}${run.error ? `: ${run.error}` : ""}`);
+    }
+    for (const e of run.summary.errors.slice(0, 5)) {
+      const where = e.nodeType ?? e.nodeId ?? "workflow";
+      issues.push(`Node ${where}: ${e.message.replace(/\s+/g, " ").slice(0, 200)}`);
+    }
+  }
+  for (const invocation of report.invocations) {
+    if (invocation.timedOutMs == null) continue;
+    issues.push(
+      `Operation "${invocation.operationId}" did not finish within its ${invocation.timedOutMs}ms timeout — the app would show it still running.`
+    );
+  }
+  // A picker over an unseeded collection renders empty, and no interaction can
+  // pick from it — worth saying, but only the script knows what belongs there.
+  for (const resource of report.resources) {
+    if (resource.seeded) continue;
+    const shown = report.widgets
+      .filter((w) => w.resourceBindingId === resource.id)
+      .map((w) => `${w.type} "${w.id}"`);
+    if (shown.length === 0) continue;
+    warnings.push(
+      `${shown.join(", ")} shows resource binding "${resource.id}", which nothing seeded — the collection is empty. Seed it with a seedResource step or a "resource:${resource.id}" param.`
+    );
+  }
+
+  const ranOperations = new Set(report.invocations.map((i) => i.operationId));
+  if (ranWorkflow && report.runs.length > 0 && report.runs.every((r) => r.ok)) {
+    const conditional = conditionalNodeIds(graph);
+    const byName = nodeIdsByName(graph);
+    const nodeIds = new Set(
+      graph.nodes.map((n) => (typeof n.id === "string" ? n.id : "")).filter(Boolean)
+    );
+    // Variables a widget writes rather than the graph — a chat composer's
+    // conversation. Nothing the run emits fills them, so an empty one after a
+    // headless run says nothing about the wiring.
+    const uiWrittenVariables = new Set(
+      (report.spec?.widgets ?? []).flatMap((w) =>
+        w.extraBindings
+          .filter((extra) => extra.ref?.kind === "variable")
+          .map((extra) =>
+            extra.ref?.kind === "variable" ? extra.ref.variableId : ""
+          )
+      )
+    );
+    for (const w of report.widgets) {
+      if (w.bindingMode !== "read" || !w.binding || w.hasValue) continue;
+      const ref = parseBinding(w.binding);
+      if (ref?.kind === "variable" && uiWrittenVariables.has(ref.variableId)) {
+        continue;
+      }
+      if (ref?.kind === "execution") {
+        // Execution state is not a value the graph emits — an empty one only
+        // means the operation never ran, or reported no activity.
+        if (!ranOperations.has(ref.operationId)) {
+          issues.push(
+            `${w.type} "${w.id}" shows operation "${ref.operationId}" ${ref.field}, but that operation never ran.`
+          );
+        } else if (ref.field === "activity") {
+          warnings.push(
+            `${w.type} "${w.id}" shows activity, but the run reported none — only streaming agent nodes emit activity labels.`
+          );
+        }
+        continue;
+      }
+      // A legacy document binds by node name, and `parseBinding` hands the name
+      // back in the `nodeId` slot, so an unrecognised id is retried as a name.
+      const parsed = (ref && "nodeId" in ref ? ref.nodeId : null) ?? w.binding;
+      const nodeId = nodeIds.has(parsed) ? parsed : byName.get(parsed) ?? null;
+      if (nodeId && conditional.has(nodeId)) {
+        // One run takes one branch, so an empty widget here is expected. It is
+        // still worth surfacing: a branch that no input can reach looks exactly
+        // the same, and only running both branches tells them apart.
+        warnings.push(
+          `${w.type} "${w.id}" is bound to "${w.binding}", downstream of a branch that was not taken this run — run the other branch to confirm it can be reached.`
+        );
+        continue;
+      }
+      issues.push(
+        `${w.type} "${w.id}" is bound to "${w.binding}" but never received a value — check the output node emits.`
+      );
+    }
+  }
+  if (
+    ranWorkflow &&
+    report.runs.length === 0 &&
+    report.invocations.length === 0 &&
+    report.spec &&
+    report.spec.widgets.length > 0
+  ) {
+    issues.push("No interaction triggered a workflow run — the app was never executed.");
+  }
+
+  const ok = issues.length === 0;
+  const headline = ok
+    ? report.runs.length > 0
+      ? `App ran clean — ${report.runs.length} run(s), every bound widget on a taken branch received a value.`
+      : "App wiring is valid (static check only — no run executed)."
+    : `App has issues — ${issues[0]}`;
+  return { ok, headline, issues, warnings };
+}
+
+/**
+ * The graph an operation runs: one the target carries (a bundle's workflows,
+ * the host workflow), else the database.
+ */
+async function resolveOperationGraph(
+  operation: OperationBinding,
+  host: { workflowId: string | null; graph: DebugGraph; graphs: Map<string, DebugGraph> },
+  loadFromDb: AppSimulationDeps["loadFromDb"]
+): Promise<{ graph: DebugGraph | null; unavailable: string | null }> {
+  const target = operation.workflowId;
+  const carried = host.graphs.get(target);
+  if (carried) return { graph: carried, unavailable: null };
+  if (!target || target === "self" || target === host.workflowId) {
+    return { graph: host.graph, unavailable: null };
+  }
+  try {
+    const workflow = await loadFromDb(target);
+    if (!workflow?.graph) {
+      return {
+        graph: null,
+        unavailable: `runs workflow "${target}", which is not in the local database.`
+      };
+    }
+    return { graph: workflow.graph, unavailable: null };
+  } catch (error) {
+    return {
+      graph: null,
+      unavailable: `runs workflow "${target}", which could not be loaded: ${String(error)}`
+    };
+  }
+}
+
+export async function simulateApp(
+  resolved: ResolvedAppTarget,
+  options: AppDebugOptions,
+  deps: AppSimulationDeps
+): Promise<AppDebugReport> {
+  const log = deps.onLog ?? (() => {});
+  const allowRuns = options.run ?? true;
+
+  const io = extractAppIO(resolved.graph);
+  // The target loader already parsed the document; `parseAppSpec` reports any
+  // issue it hit.
+  const document = resolved.document;
+
+  // Resolve every declared operation against a real graph before parsing the
+  // widgets: a binding into a second operation resolves against that
+  // operation's surface, not the host's.
+  const graphByOperation = new Map<string, DebugGraph>();
+  const context: AppContext = {
+    defaultOperationId: DEFAULT_OPERATION_ID,
+    operations: [],
+    variables: document?.variables ?? [],
+    resources: document?.resources.map(({ id, name, kind }) => ({ id, name, kind })) ?? []
+  };
+  const bindingByOperation = new Map<string, OperationBinding>();
+  if (document) {
+    for (const binding of documentOperations(document)) {
+      const { graph, unavailable } = await resolveOperationGraph(
+        binding,
+        {
+          workflowId: resolved.info.workflowId,
+          graph: resolved.graph,
+          graphs: resolved.graphs
+        },
+        deps.loadFromDb
+      );
+      const operationIO: AppIO | null =
+        graph === resolved.graph ? io : graph ? extractAppIO(graph) : null;
+      if (graph) graphByOperation.set(binding.id, graph);
+      bindingByOperation.set(binding.id, binding);
+      context.operations.push(operationSpec(binding, operationIO, unavailable));
+    }
+    context.defaultOperationId =
+      context.operations.find((op) => op.id === DEFAULT_OPERATION_ID)?.id ??
+      context.operations[0]?.id ??
+      DEFAULT_OPERATION_ID;
+  }
+
+  const { spec, issues: parseIssues, warnings: parseWarnings } = parseAppSpec(
+    resolved.appDoc,
+    io,
+    document ? context : undefined,
+    { document, issue: resolved.issue }
+  );
+  const validation = spec
+    ? validateApp(spec, io, context)
+    : { errors: [], warnings: [] };
+  validation.errors = [...parseIssues, ...validation.errors];
+  validation.warnings = [...parseWarnings, ...validation.warnings];
+
+  const report: AppDebugReport = {
+    generatedAt: new Date().toISOString(),
+    target: resolved.info,
+    app: spec
+      ? {
+          version: spec.version,
+          // A document keeps its title in the Puck root; an application row
+          // keeps it as the row's name.
+          title: spec.title ?? resolved.appName,
+          widgetCount: spec.widgets.length
+        }
+      : null,
+    spec,
+    io: {
+      inputs: io.inputs.map((i) => i.name),
+      outputs: io.outputs.map((o) => o.name),
+      variables: io.variables
+    },
+    validation,
+    interactions: [],
+    runs: [],
+    values: {},
+    variables: {},
+    invocations: [],
+    activity: [],
+    widgets: [],
+    resources: [],
+    notSimulated: [...NOT_SIMULATED],
+    verdict: { ok: false, headline: "", issues: [] },
+    bundleDir: null
+  };
+
+  /** Verdict issues the condition simulation found, folded in at the end. */
+  const conditionIssues: string[] = [];
+
+  if (spec) {
+    const runs: ServerRunReport[] = report.runs;
+    const runWorkflow = async (
+      operationId: string,
+      graph: DebugGraph,
+      workflowId: string | null,
+      timeoutMs: number | null,
+      params: Record<string, unknown>
+    ) => {
+      log(`Running operation "${operationId}"…`);
+      const outcome = await deps.runOnServer({
+        graph,
+        workflowId,
+        params,
+        ...(timeoutMs != null ? { timeoutMs } : {})
+      });
+      // The slot is claimed only once the run settles, so a run the harness
+      // timed out on never takes an index from the one that follows it.
+      const runIndex = runs.length;
+      const messagesFile = await deps.onRunMessages?.(runIndex, outcome.rawMessages);
+      if (messagesFile) outcome.report.messagesFile = messagesFile;
+      runs.push(outcome.report);
+      log(
+        `Run ${runIndex + 1}: ${outcome.report.status} · ${outcome.report.summary.counts.errored} node error(s)`
+      );
+      return {
+        messages: outcome.rawMessages as unknown as ReadonlyArray<
+          Record<string, unknown>
+        >,
+        runIndex
+      };
+    };
+
+    const scope = bindingScopeFor(io, context);
+    const operations: HeadlessOperationInit[] = context.operations.map((op) => {
+      const binding = bindingByOperation.get(op.id) as OperationBinding;
+      const graph = graphByOperation.get(op.id) ?? null;
+      const operationIO = op.io ?? { inputs: [], outputs: [], variables: [], nodeIds: [] };
+      const defaults: Record<string, unknown> = {};
+      for (const input of operationIO.inputs) {
+        if (input.defaultValue === undefined) continue;
+        defaults[stateKey({ kind: "input", operationId: op.id, nodeId: input.nodeId })] =
+          input.defaultValue;
+      }
+      const timeoutMs = effectiveTimeoutMs(op.timeoutMs, options.timeoutMs);
+      return {
+        binding,
+        outputKeyByNodeId: new Map(
+          operationIO.outputs.map((o) => [
+            o.nodeId,
+            stateKey({ kind: "output", operationId: op.id, nodeId: o.nodeId })
+          ])
+        ),
+        inputNodeIds: operationIO.inputs.map((i) => i.nodeId),
+        inputNameByNodeId: new Map(operationIO.inputs.map((i) => [i.nodeId, i.name])),
+        defaults,
+        ...(graph
+          ? {
+              runWorkflow: (params: Record<string, unknown>) =>
+                runWorkflow(
+                  op.id,
+                  graph,
+                  // A bundle's operations name bundle-local keys, so there is
+                  // no workflow id to attribute the run to.
+                  resolved.operationsReferenceKeys
+                    ? null
+                    : graph === resolved.graph
+                      ? resolved.info.workflowId
+                      : op.workflowId,
+                  timeoutMs,
+                  params
+                )
+            }
+          : {})
+      };
+    });
+
+    const runtime = new HeadlessAppRuntime({
+      operations,
+      defaultOperationId: context.defaultOperationId,
+      variables: context.variables,
+      resources: document?.resources ?? [],
+      scope,
+      widgets: spec.widgets.map((w) => ({
+        id: w.id,
+        ...(w.visibleWhen ? { visibleWhen: w.visibleWhen } : {}),
+        ...(w.disabledWhen ? { disabledWhen: w.disabledWhen } : {}),
+        ...(w.format ? { format: w.format } : {})
+      })),
+      ...(options.timeoutMs != null ? { timeoutMs: options.timeoutMs } : {})
+    });
+
+    /** Write a value addressed by name (a param, or an `--interact` set step). */
+    const writeByName = (
+      name: string,
+      value: unknown,
+      operationId = context.defaultOperationId
+    ): boolean => {
+      const ref =
+        resolveBinding(name, scope, "write", operationId) ??
+        resolveBinding(name, scope, "read", operationId);
+      if (!ref) return false;
+      runtime.write(ref, value);
+      return true;
+    };
+
+    /** Seed one collection, reporting why a bad seed did nothing. */
+    const seedResource = (
+      resourceBindingId: string,
+      items: unknown
+    ): string | null => {
+      if (!Array.isArray(items)) {
+        return `Resource seed for "${resourceBindingId}" is not an array of items.`;
+      }
+      const seeds: SeedResourceItem[] = [];
+      for (const item of items) {
+        const seed = item as SeedResourceItem | null;
+        if (!seed || typeof seed.id !== "string" || seed.id.length === 0) {
+          return `Resource seed for "${resourceBindingId}" has an item without an "id".`;
+        }
+        seeds.push(seed);
+      }
+      try {
+        runtime.seedResource(resourceBindingId, seeds);
+        return null;
+      } catch (error) {
+        return (error as Error).message;
+      }
+    };
+
+    for (const [key, value] of Object.entries({
+      ...resolved.fileParams,
+      ...(options.params ?? {})
+    })) {
+      // `--params '{"resource:docs": [...]}'` seeds a collection; every other
+      // key writes a reactive value.
+      if (key.startsWith("resource:")) {
+        const failure = seedResource(key.slice("resource:".length), value);
+        if (failure) report.validation.warnings.push(failure);
+        continue;
+      }
+      if (!writeByName(key, value)) {
+        report.validation.warnings.push(
+          `Param "${key}" matches no input, output, or variable in the workflow.`
+        );
+      }
+    }
+
+    /** Dispatch one action, tracking which run it produced. */
+    const dispatch = async (
+      record: InteractionRecord,
+      action: Parameters<HeadlessAppRuntime["dispatch"]>[0]
+    ): Promise<void> => {
+      const before = runtime.invocations.length;
+      try {
+        await runtime.dispatch(action);
+      } catch (error) {
+        record.error = record.error ?? (error as Error).message;
+        return;
+      }
+      const started = runtime.invocations.slice(before);
+      const withRun = started.find((i) => i.runIndex !== null);
+      if (withRun) record.runIndex = withRun.runIndex;
+      if (action.kind === "run") {
+        record.error = record.error ?? runtime.errorFor(action.operationId);
+      }
+    };
+
+    const steps = options.interact ?? defaultInteractions(spec);
+    for (const step of steps) {
+      const record: InteractionRecord = {
+        step: describeStep(step),
+        actions: [],
+        runIndex: null,
+        error: null
+      };
+      report.interactions.push(record);
+
+      if ("set" in step) {
+        if (writeByName(step.set.key, step.set.value, step.set.operationId)) {
+          record.actions.push(`set ${step.set.key}`);
+        } else {
+          record.error = `"${step.set.key}" matches no input, output, or variable.`;
+        }
+        continue;
+      }
+
+      if ("seedResource" in step) {
+        const { id, items } = step.seedResource;
+        const failure = seedResource(id, items);
+        if (failure) record.error = failure;
+        else record.actions.push(`seedResource ${id} (${items.length} item(s))`);
+        continue;
+      }
+
+      if ("run" in step) {
+        if (!allowRuns) {
+          record.actions.push("run (skipped — --no-run)");
+          continue;
+        }
+        record.actions.push(`run ${step.run}`);
+        await dispatch(record, { kind: "run", operationId: step.run });
+        continue;
+      }
+
+      if ("cancel" in step) {
+        const cancelled = runtime.cancel(step.cancel);
+        record.actions.push(
+          cancelled.length > 0
+            ? `cancel ${step.cancel} (${cancelled.join(", ")})`
+            : `cancel ${step.cancel} (nothing running)`
+        );
+        continue;
+      }
+
+      const trigger = "click" in step ? "click" : "change";
+      const found = findWidget(spec, "click" in step ? step.click : step.change);
+      if (typeof found === "string") {
+        record.error = found;
+        continue;
+      }
+      // A `set` step writes state directly and models the runtime; a click or a
+      // change models a user, and a user cannot touch what the app hides or
+      // disables.
+      const gate = runtime.widgetState(found.id);
+      const verb = trigger === "click" ? "clicked" : "changed";
+      const named = found.label ?? found.id;
+      if (gate && !gate.visible) {
+        record.error = `${verb} "${named}" while hidden by \`${gate.visibleWhen ?? "visibleWhen"}\`.`;
+        continue;
+      }
+      if (gate?.disabled) {
+        record.error = `${verb} "${named}" while disabled by \`${gate.disabledWhen ?? "disabledWhen"}\`.`;
+        continue;
+      }
+      if ("change" in step && step.value !== undefined) {
+        if (found.ref) runtime.write(found.ref, step.value);
+        record.actions.push(`set ${found.stateKey ?? found.id}`);
+      }
+      // A run from a bound write widget carries its binding; the web engine
+      // would run just that input's downstream subgraph — headless, both paths
+      // are a full authoritative run.
+      const from =
+        found.bindingMode === "write" ? found.canonicalBinding ?? undefined : undefined;
+      const eventCtx = {
+        defaultOperationId: context.defaultOperationId,
+        resolveVariableId: (key: string | undefined) => {
+          const ref = resolveBinding(key, scope, "read");
+          return ref?.kind === "variable" ? ref.variableId : null;
+        },
+        from
+      };
+      for (const event of found.events) {
+        if (event.trigger !== trigger) continue;
+        const action = eventToAction(event, eventCtx);
+        if (!action) {
+          record.error =
+            record.error ?? `event "${event.kind}" is incomplete — nothing to dispatch.`;
+          continue;
+        }
+        if (action.kind === "run" && !allowRuns) {
+          record.actions.push("run (skipped — --no-run)");
+          continue;
+        }
+        record.actions.push(
+          action.kind === "setVariable" || action.kind === "toggleVariable"
+            ? `${action.kind} ${action.variableId}`
+            : action.kind === "run" || action.kind === "cancel"
+              ? `${action.kind} ${action.operationId ?? context.defaultOperationId}`
+              : action.kind
+        );
+        await dispatch(record, action);
+      }
+      if (record.actions.length === 0) {
+        record.error = `widget has no "${trigger}" events to fire.`;
+      }
+    }
+
+    // The report reads better keyed by the names an author recognizes than by
+    // the runtime's namespaced state keys. A second operation prefixes its
+    // names so two operations over one workflow stay apart.
+    const values: Record<string, unknown> = {};
+    for (const operation of context.operations) {
+      const prefix =
+        operation.id === context.defaultOperationId ? "" : `${operation.id}.`;
+      for (const input of operation.io?.inputs ?? []) {
+        const value = runtime.read({
+          kind: "input",
+          operationId: operation.id,
+          nodeId: input.nodeId
+        });
+        if (value !== undefined) values[`${prefix}${input.name}`] = value;
+      }
+      for (const output of operation.io?.outputs ?? []) {
+        const value = runtime.read({
+          kind: "output",
+          operationId: operation.id,
+          nodeId: output.nodeId
+        });
+        if (value !== undefined) values[`${prefix}${output.name}`] = value;
+      }
+    }
+    const variables: Record<string, unknown> = {};
+    for (const [variableId, value] of Object.entries(runtime.state.variables)) {
+      if (value === undefined) continue;
+      values[variableId] = value;
+      variables[variableId] = value;
+    }
+    report.values = previewValue(values) as Record<string, unknown>;
+    report.variables = previewValue(variables) as Record<string, unknown>;
+
+    report.invocations = runtime.invocations.map((invocation) => ({
+      id: invocation.id,
+      operationId: invocation.operationId,
+      status: runtime.state.invocations[invocation.id]?.status ?? "unknown",
+      decision: invocation.decision,
+      decisionTargets: invocation.decisionTargets,
+      runIndex: invocation.runIndex,
+      timedOutMs: invocation.timedOutMs,
+      error: runtime.state.invocations[invocation.id]?.error ?? null,
+      activity: invocation.activity
+    }));
+    report.activity = [...runtime.activity];
+
+    report.resources = runtime.resourceBindingList.map((binding) => {
+      const provider = runtime.resourceProvider(binding.id);
+      const items = provider?.list() ?? [];
+      return {
+        id: binding.id,
+        kind: binding.kind,
+        seeded: provider !== null,
+        items: items.map((item) => ({
+          id: item.ref.id,
+          name: item.name,
+          ...(item.ref.revision !== undefined ? { revision: item.ref.revision } : {})
+        })),
+        selected: provider?.selected()?.id ?? null,
+        commands: runtime.resourceCommands
+          .filter((c) => c.resourceBindingId === binding.id)
+          .map((c) => c.command)
+      };
+    });
+
+    report.widgets = spec.widgets
+      .filter((w) => w.bindingMode !== "layout")
+      .map((w) => {
+        // A resource widget shows a collection, not a state value — report what
+        // its picker or gallery would list.
+        if (w.resourceBindingId) {
+          const items = runtime.resourceProvider(w.resourceBindingId)?.list() ?? [];
+          const state = runtime.widgetState(w.id);
+          return {
+            id: w.id,
+            type: w.type,
+            bindingMode: w.bindingMode,
+            binding: w.binding,
+            stateKey: w.stateKey,
+            resourceBindingId: w.resourceBindingId,
+            value: previewValue(
+              items.map((item) => ({ id: item.ref.id, name: item.name }))
+            ),
+            display: null,
+            hasValue: items.length > 0,
+            visible: state?.visible ?? true,
+            disabled: state?.disabled ?? false
+          };
+        }
+        const value = runtime.read(w.ref);
+        // A `format` template is what the widget actually shows, so it — not
+        // the raw value — decides whether anything reached the screen.
+        const display = runtime.display(w.id);
+        const state = runtime.widgetState(w.id);
+        return {
+          id: w.id,
+          type: w.type,
+          bindingMode: w.bindingMode,
+          binding: w.binding,
+          stateKey: w.stateKey,
+          value: previewValue(value),
+          display,
+          hasValue: display !== null ? display.length > 0 : value !== undefined,
+          visible: state?.visible ?? true,
+          disabled: state?.disabled ?? false
+        };
+      });
+
+    for (const w of spec.widgets) {
+      if (!w.events.some((e) => e.kind === "run")) continue;
+      const state = runtime.widgetState(w.id);
+      if (!state || state.everReachable) continue;
+      const blocking = !state.visible
+        ? `visibleWhen (\`${state.visibleWhen ?? ""}\`) never held`
+        : `disabledWhen (\`${state.disabledWhen ?? ""}\`) always held`;
+      conditionIssues.push(
+        `${w.type} "${w.label ?? w.id}" runs the app, but its ${blocking} — no interaction could start a run.`
+      );
+    }
+  }
+
+  report.verdict = buildAppVerdict(report, allowRuns, resolved.graph, conditionIssues);
+
+  return report;
+}
