+++ packages/execution/src/debug/types.ts
@@ -9,6 +9,7 @@
  */
 
 import type { Intervention } from "@nodetool-ai/protocol";
+import type { SupervisedRunSummary } from "../supervisor.js";
 
 /** A single value emitted by a node (output_update / generation_complete). */
 export interface NodeOutput {
@@ -97,4 +98,80 @@ export interface RunVerdict {
   ok: boolean;
   headline: string;
   issues: string[];
+}
+
+// ─── Target and run-report vocabulary ────────────────────────────────────────
+// The CLI debug harness and the app simulator both describe a target and a
+// finished server run, so the shapes live here rather than in either host.
+
+/** A workflow graph in kernel/runner shape (properties, not data). */
+export interface DebugGraph {
+  nodes: Array<Record<string, unknown>>;
+  edges: Array<Record<string, unknown>>;
+}
+
+/** How the workflow target was provided and what it resolved to. */
+export interface DebugTargetInfo {
+  /** Original CLI argument (id, path/to/file.json, or path/to/file.ts). */
+  ref: string;
+  /**
+   * How `ref` was interpreted. `application` and `bundle` only occur for app
+   * targets: an application row read from the database, and an
+   * `ApplicationBundle` JSON file.
+   */
+  source: "id" | "json" | "dsl" | "application" | "bundle";
+  /** Workflow id when known (DB id, or `id`/`workflow_id` field in a file). */
+  workflowId: string | null;
+  nodeCount: number;
+  edgeCount: number;
+}
+
+/** Rolled-up OpenTelemetry spans for the run. */
+export interface TraceSummary {
+  spanCount: number;
+  /** Wall-clock span over all spans (max end − min start), ms. */
+  totalDurationMs: number | null;
+  tokens: { input: number; output: number; total: number };
+  costUsd: number;
+  /** Count + total self-time by span name (llm.chat, node.process, …). */
+  byName: Record<string, { count: number; totalDurationMs: number }>;
+  /** Slowest spans, descending. */
+  slowest: Array<{ name: string; durationMs: number; status: string }>;
+}
+
+export interface ServerRunReport {
+  surface: "server";
+  /** True when the job reached the `completed` status. */
+  ok: boolean;
+  status: string;
+  error: string | null;
+  durationMs: number;
+  summary: ExecutionSummary;
+  trace: TraceSummary | null;
+  /** Bundle-relative path to the raw messages JSONL. */
+  messagesFile?: string;
+  /** Bundle-relative path to the raw trace JSONL. */
+  traceFile?: string;
+  /**
+   * Present only on a supervised run: who supervised it and the rollup of what
+   * they decided. The decisions themselves are `summary.interventions`.
+   */
+  supervised?: {
+    provider: string;
+    model: string;
+    summary: SupervisedRunSummary;
+  };
+}
+
+/** A verdict that can also carry non-fatal observations. */
+export interface DebugVerdict {
+  ok: boolean;
+  headline: string;
+  /** Human-readable problems found, ordered most-actionable first. */
+  issues: string[];
+  /**
+   * Things worth looking at that are not failures — a run cannot decide them
+   * either way. Unlike `issues`, these do not clear `ok`.
+   */
+  warnings?: string[];
 }
